Books Room At Specified Date/Time
Given any email request it is able to process that request and
book a room at that specified date and time. Automates the
whole process from login to confirmation. However, need to
add error handling for when the booking fails. Created the
frame work for sending error emails when something fails.